move instances into separate trait
