+++ core/src/main/scala/cats/Cartesian.scala
@@ -16,14 +16,4 @@ import simulacrum.typeclass
   def product[A, B](fa: F[A], fb: F[B]): F[(A, B)]
 }
 
-object Cartesian extends CartesianArityFunctions with KernelCartesianInstances
-
-/**
- * Cartesian instances for types that are housed in Kernel and therefore
- * can't have instances for Cats type classes in their companion objects.
- */
-private[cats] sealed trait KernelCartesianInstances {
-  implicit val catsInvariantSemigroup: Cartesian[Semigroup] = InvariantMonoidal.catsInvariantMonoidalSemigroup
-  implicit val catsInvariantMonoid: Cartesian[Monoid] = InvariantMonoidal.catsInvariantMonoidalMonoid
-  implicit val catsCartesianEq: Cartesian[Eq] = ContravariantCartesian.catsContravariantCartesianEq
-}
+object Cartesian extends CartesianArityFunctions
