+++ core/src/main/scala/cats/functor/Contravariant.scala
@@ -28,32 +28,3 @@ import simulacrum.typeclass
       val G = Functor[G]
     }
 }
-
-object Contravariant extends KernelContravariantInstances
-
-/**
- * Convariant instances for types that are housed in cats.kernel and therefore
- * can't have instances for this type class in their companion objects.
- */
-private[functor] sealed trait KernelContravariantInstances {
-  implicit def catsFunctorContravariantForEq: Contravariant[Eq] =
-    ContravariantCartesian.catsContravariantCartesianEq
-
-  implicit val catsFunctorContravariantForPartialOrder: Contravariant[PartialOrder] =
-    new Contravariant[PartialOrder] {
-      /** Derive a `PartialOrder` for `B` given a `PartialOrder[A]` and a function `B => A`.
-       *
-       * Note: resulting instances are law-abiding only when the functions used are injective (represent a one-to-one mapping)
-       */
-      def contramap[A, B](fa: PartialOrder[A])(f: B => A): PartialOrder[B] = fa.on(f)
-    }
-
-  implicit val catsFunctorContravariantForOrder: Contravariant[Order] =
-    new Contravariant[Order] {
-      /** Derive an `Order` for `B` given an `Order[A]` and a function `B => A`.
-       *
-       * Note: resulting instances are law-abiding only when the functions used are injective (represent a one-to-one mapping)
-       */
-      def contramap[A, B](fa: Order[A])(f: B => A): Order[B] = fa.on(f)
-    }
-}
