+++ laws/src/main/scala/cats/laws/discipline/FlatMapTests.scala
@@ -2,6 +2,8 @@ package cats
 package laws
 package discipline
 
+import cats.instances.all._
+
 import cats.laws.discipline.CartesianTests.Isomorphisms
 import org.scalacheck.{Arbitrary, Cogen, Prop}
 import Prop._
